Disable curl's URL globbing parser
Users may not expect URLs that include [] and {} character pairs to be
interpolated by curl's globbing parser when they are expecting the
behavior of a simple download tool.  This change adds the "--globoff"
curl long option to the list of curl options applied by default, to
try to provide default behavior that is as unsurprising as possible.

Users who want the globbing behavior still have the option of enabling
it with the "--no-globoff" curl option, passed in via wcurl's
-o/--opts=... mechanism.