Improve wcurl version comparison mechanism
We've been having a some problems with version comparison in the
script.  The approach we've taken so far is to decompose the current
curl version into two variables (major and minor), and then perform
comparisons against these components separately.  It works, but it's
confusing.

Another possible approach would be to use a C-style version
normalization (basically a printf "%02d%02d%02d") and then perform
comparisons against the versions we want.  The problem is that these
versions need also to be normalized, which can be confusing as well.

I decided to implement the second approach but abstract it as a simple
function that can take a regular version string like "8.16.0" as well as
a comparison operator that will then be passed onto to "test".  This
reads much nicer and abstracts the complexities of version
normalization away.  Unfortunately due to the limitations of shell
scripting it's not easy to deduplicate code in this scenario, but that
should be OK for now.

Signed-off-by: Sergio Durigan Junior <sergiodj@sergiodj.net>