tests/tests.sh: Fix ROOTDIR definition
Back in ea192a3c2b057ab4b0a5ee79d541656542667bfc the definition of the
ROOTPATH variable was changed to fix some shellcheck errors.  The
problem is that the new definition wrongly causes the variable to have
an empty value, effectively making our tests to run against the
installed wcurl version instead of the version we're actually
developing.

This commit fixes the problem by simplifying the ROOTPATH definition
again while making it explicit that CDPATH's value should be empty.

Signed-off-by: Sergio Durigan Junior <sergiodj@sergiodj.net>